Fix tools/valgrind/chrome_tests.sh to always use --log-file option so fork() doesn't corrupt log files; lets --generate_suppressions work even with the hacky valgrind fix that makes anything but --log-file crash on child of fork().
Also avoid killing user desktop session if zygote dies!

BUG=none, but related to fix for http://crbug.com/15771
TEST=patch valgrind with fork workaround; sh tools/valgrind/chrome_tests.sh --generate_suppressions -t ui logs you out on linux without this.

Review URL: http://codereview.chromium.org/155130

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20024 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/process_util_posix.cc
@@ -64,6 +64,12 @@ ProcessId GetProcId(ProcessHandle process) {
 // entry structure.  Ignores specified exit_code; posix can't force that.
 // Returns true if this is successful, false otherwise.
 bool KillProcess(ProcessHandle process_id, int exit_code, bool wait) {
+  DCHECK(process_id > 1);
+  if (process_id <= 1) {
+    LOG(ERROR) << "tried to kill process_id " << process_id;
+    return false;
+  }
+
   bool result = kill(process_id, SIGTERM) == 0;
 
   if (result && wait) {
